allow system probe to build and load config on darwin
