+++ cmd/system-probe/config/config.go
@@ -48,11 +48,6 @@ func New(configPath string) (*types.Config, error) {
 }
 
 func newSysprobeConfig(configPath string) (*types.Config, error) {
-	// System probe is not supported on darwin, so we should fail gracefully in this case.
-	if runtime.GOOS == "darwin" {
-		return &types.Config{}, nil
-	}
-
 	aconfig.SystemProbe.SetConfigName("system-probe")
 	// set the paths where a config file is expected
 	if len(configPath) != 0 {
