+++ cmd/system-probe/config/config_darwin.go
@@ -3,21 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !unix && !windows
+//go:build darwin
 
 package config
-
-import "fmt"
-
-const (
-	defaultConfigDir          = ""
-	defaultSystemProbeAddress = ""
-)
-
-// ValidateSocketAddress validates that the sysprobe socket config option is of the correct format.
-func ValidateSocketAddress(sockPath string) error { //nolint:revive // TODO fix revive unused-parameter
-	return fmt.Errorf("system-probe unsupported")
-}
 
 // ProcessEventDataStreamSupported returns true if process event data stream is supported
 func ProcessEventDataStreamSupported() bool {
