+++ cmd/system-probe/config/config_linux.go
@@ -8,27 +8,9 @@
 package config
 
 import (
-	"fmt"
-	"path/filepath"
-
 	ebpfkernel "github.com/DataDog/datadog-agent/pkg/security/ebpf/kernel"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
-
-const (
-	// defaultSystemProbeAddress is the default unix socket path to be used for connecting to the system probe
-	defaultSystemProbeAddress = "/opt/datadog-agent/run/sysprobe.sock"
-
-	defaultConfigDir = "/etc/datadog-agent"
-)
-
-// ValidateSocketAddress validates that the sysprobe socket config option is of the correct format.
-func ValidateSocketAddress(sockPath string) error {
-	if !filepath.IsAbs(sockPath) {
-		return fmt.Errorf("socket path must be an absolute file path: `%s`", sockPath)
-	}
-	return nil
-}
 
 // ProcessEventDataStreamSupported returns true if process event data stream is supported
 func ProcessEventDataStreamSupported() bool {
