+++ cmd/system-probe/main.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux
+//go:build unix
 
 // Package main is the entrypoint for system-probe process
 package main
